[guide] move guide to core
